test: remove duplicate and redundant test files
- Remove 8 duplicate test files (~3,060 lines)
- Simplify over-engineered randomPassword test (216→54 lines)
- Consolidate auth tests (keep comprehensive version)
- Consolidate chat message tests (remove 3 duplicate GET tests)
- Remove duplicate utility tests (slugify, randomPassword, enum, utils)
- Remove duplicate workspace-member-queries test

All tests still passing (169/169).
Test coverage maintained with cleaner, more maintainable test suite.
