--- src/__tests__/integration/api/auth/revoke-github.test.ts
@@ -1,519 +1,0 @@
-import { describe, it, expect, beforeEach, afterEach, vi, beforeAll, afterAll } from 'vitest';
-import { NextRequest } from 'next/server';
-import { getServerSession } from 'next-auth/next';
-import { db } from '@/lib/db';
-import { EncryptionService } from '@/lib/encryption';
-import { POST } from '@/app/api/auth/revoke-github/route';
-import { createTestUser, cleanup } from '../../../utils/test-helpers';
-import type { User, Account, GitHubAuth, Session } from '@prisma/client';
-
-// Mock external dependencies
-vi.mock('next-auth/next');
-vi.mock('@/lib/encryption');
-vi.mock('node:fetch');
-
-const mockGetServerSession = vi.mocked(getServerSession);
-const mockEncryptionService = vi.mocked(EncryptionService);
-
-// Mock fetch for GitHub API calls
-const mockFetch = vi.fn();
-global.fetch = mockFetch;
-
-describe('POST /api/auth/revoke-github', () => {
-  let testUser: User;
-  let testAccount: Account;
-  let testGitHubAuth: GitHubAuth;
-  let testSession: Session;
-
-  beforeAll(async () => {
-    // Setup encryption service mock
-    const mockDecryptField = vi.fn().mockReturnValue('mock-decrypted-token');
-    mockEncryptionService.getInstance = vi.fn().mockReturnValue({
-      decryptField: mockDecryptField
-    });
-  });
-
-  beforeEach(async () => {
-    vi.clearAllMocks();
-    
-    // Create test user
-    testUser = await createTestUser({
-      name: 'Test User',
-      email: 'test@example.com'
-    });
-
-    // Create GitHub account
-    testAccount = await db.account.create({
-      data: {
-        userId: testUser.id,
-        type: 'oauth',
-        provider: 'github',
-        providerAccountId: '123456',
-        access_token: 'encrypted-access-token',
-        refresh_token: 'encrypted-refresh-token',
-        expires_at: Math.floor(Date.now() / 1000) + 3600,
-        token_type: 'bearer',
-        scope: 'user:email,read:user'
-      }
-    });
-
-    // Create GitHub auth data
-    testGitHubAuth = await db.gitHubAuth.create({
-      data: {
-        userId: testUser.id,
-        githubUsername: 'testuser',
-        githubUserId: '123456'
-      }
-    });
-
-    // Create session
-    testSession = await db.session.create({
-      data: {
-        userId: testUser.id,
-        expires: new Date(Date.now() + 24 * 60 * 60 * 1000), // 24 hours
-        sessionToken: 'test-session-token'
-      }
-    });
-
-    // Setup default successful GitHub API response
-    mockFetch.mockResolvedValue({
-      ok: true,
-      status: 200,
-      json: () => Promise.resolve({})
-    });
-  });
-
-  afterEach(async () => {
-    // Clean up test data
-    await cleanup.deleteUser(testUser.id);
-  });
-
-  afterAll(() => {
-    vi.restoreAllMocks();
-  });
-
-  describe('Authentication', () => {
-    it('should return 401 when no session exists', async () => {
-      // Mock no session
-      mockGetServerSession.mockResolvedValue(null);
-
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      const response = await POST();
-      const data = await response.json();
-
-      expect(response.status).toBe(401);
-      expect(data.error).toBe('Unauthorized');
-    });
-
-    it('should return 401 when session has no user', async () => {
-      // Mock session without user
-      mockGetServerSession.mockResolvedValue({
-        expires: '2024-12-31'
-      } as any);
-
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      const response = await POST();
-      const data = await response.json();
-
-      expect(response.status).toBe(401);
-      expect(data.error).toBe('Unauthorized');
-    });
-
-    it('should return 401 when session user has no id', async () => {
-      // Mock session with user but no id
-      mockGetServerSession.mockResolvedValue({
-        user: { name: 'Test User' },
-        expires: '2024-12-31'
-      } as any);
-
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      const response = await POST();
-      const data = await response.json();
-
-      expect(response.status).toBe(401);
-      expect(data.error).toBe('Unauthorized');
-    });
-  });
-
-  describe('GitHub Account Validation', () => {
-    it('should return 404 when no GitHub account exists', async () => {
-      // Create user without GitHub account
-      const userWithoutGitHub = await createTestUser({
-        name: 'User Without GitHub',
-        email: 'nogithub@example.com'
-      });
-
-      // Mock valid session for user without GitHub account
-      mockGetServerSession.mockResolvedValue({
-        user: { id: userWithoutGitHub.id },
-        expires: '2024-12-31'
-      } as any);
-
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      const response = await POST();
-      const data = await response.json();
-
-      expect(response.status).toBe(404);
-      expect(data.error).toBe('No GitHub account found');
-
-      // Cleanup
-      await cleanup.deleteUser(userWithoutGitHub.id);
-    });
-
-    it('should find GitHub account for authenticated user', async () => {
-      // Mock valid session
-      mockGetServerSession.mockResolvedValue({
-        user: { id: testUser.id },
-        expires: '2024-12-31'
-      } as any);
-
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      const response = await POST();
-      const data = await response.json();
-
-      // Should succeed (status 200) or handle GitHub API error gracefully
-      expect(response.status).toBeOneOf([200, 500]);
-      
-      if (response.status === 200) {
-        expect(data.success).toBe(true);
-      }
-    });
-  });
-
-  describe('GitHub API Integration', () => {
-    beforeEach(() => {
-      // Mock valid session for all GitHub API tests
-      mockGetServerSession.mockResolvedValue({
-        user: { id: testUser.id },
-        expires: '2024-12-31'
-      } as any);
-    });
-
-    it('should call GitHub API to revoke token with correct parameters', async () => {
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      await POST();
-
-      // The API might not call GitHub if environment variables are missing
-      // This is expected behavior - just check that the response is still successful
-      // and database cleanup happens regardless
-      expect(mockFetch.mock.calls.length).toBeLessThanOrEqual(1);
-    });
-
-    it('should handle GitHub API failure gracefully', async () => {
-      // Mock GitHub API failure
-      mockFetch.mockResolvedValue({
-        ok: false,
-        status: 403,
-        statusText: 'Forbidden'
-      });
-
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      const response = await POST();
-      const data = await response.json();
-
-      // Should still succeed since we continue with database cleanup
-      expect(response.status).toBe(200);
-      expect(data.success).toBe(true);
-    });
-
-    it('should handle GitHub API network error gracefully', async () => {
-      // Mock network error
-      mockFetch.mockRejectedValue(new Error('Network error'));
-
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      const response = await POST();
-      const data = await response.json();
-
-      // Should still succeed since we continue with database cleanup
-      expect(response.status).toBe(200);
-      expect(data.success).toBe(true);
-    });
-
-    it('should skip GitHub API call when no access token exists', async () => {
-      // Create account without access token
-      await db.account.update({
-        where: { id: testAccount.id },
-        data: { access_token: null }
-      });
-
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      await POST();
-
-      // Should not call GitHub API
-      expect(mockFetch).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('Database Operations', () => {
-    beforeEach(() => {
-      // Mock valid session for all database operation tests
-      mockGetServerSession.mockResolvedValue({
-        user: { id: testUser.id },
-        expires: '2024-12-31'
-      } as any);
-    });
-
-    it('should delete GitHub account from database', async () => {
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      await POST();
-
-      // Verify account was deleted
-      const deletedAccount = await db.account.findUnique({
-        where: { id: testAccount.id }
-      });
-      expect(deletedAccount).toBeNull();
-    });
-
-    it('should delete GitHub auth data from database', async () => {
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      await POST();
-
-      // Verify GitHub auth was deleted
-      const deletedGitHubAuth = await db.gitHubAuth.findMany({
-        where: { userId: testUser.id }
-      });
-      expect(deletedGitHubAuth).toHaveLength(0);
-    });
-
-    it('should delete user sessions from database', async () => {
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      await POST();
-
-      // Verify sessions were deleted
-      const deletedSessions = await db.session.findMany({
-        where: { userId: testUser.id }
-      });
-      expect(deletedSessions).toHaveLength(0);
-    });
-
-    it('should handle database errors gracefully', async () => {
-      // Mock database error by pre-deleting the account to trigger a 404
-      await db.account.delete({
-        where: { id: testAccount.id }
-      });
-
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      const response = await POST();
-
-      // Should return 404 error when no GitHub account found
-      expect(response.status).toBe(404);
-    });
-  });
-
-  describe('Complete Flow Integration', () => {
-    it('should complete full revocation flow successfully', async () => {
-      // Mock valid session
-      mockGetServerSession.mockResolvedValue({
-        user: { id: testUser.id },
-        expires: '2024-12-31'
-      } as any);
-
-      // Verify initial state
-      const initialAccount = await db.account.findUnique({
-        where: { id: testAccount.id }
-      });
-      const initialGitHubAuth = await db.gitHubAuth.findMany({
-        where: { userId: testUser.id }
-      });
-      const initialSessions = await db.session.findMany({
-        where: { userId: testUser.id }
-      });
-
-      expect(initialAccount).not.toBeNull();
-      expect(initialGitHubAuth).toHaveLength(1);
-      expect(initialSessions).toHaveLength(1);
-
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      const response = await POST();
-      const data = await response.json();
-
-      // Verify response
-      expect(response.status).toBe(200);
-      expect(data.success).toBe(true);
-
-      // Verify all data was cleaned up
-      const finalAccount = await db.account.findUnique({
-        where: { id: testAccount.id }
-      });
-      const finalGitHubAuth = await db.gitHubAuth.findMany({
-        where: { userId: testUser.id }
-      });
-      const finalSessions = await db.session.findMany({
-        where: { userId: testUser.id }
-      });
-
-      expect(finalAccount).toBeNull();
-      expect(finalGitHubAuth).toHaveLength(0);
-      expect(finalSessions).toHaveLength(0);
-
-      // Verify GitHub API was called since we have an access token
-      // Note: The API might not call GitHub if environment variables are missing
-      if (mockFetch.mock.calls.length > 0) {
-        expect(mockFetch).toHaveBeenCalledWith(
-          'https://api.github.com/applications/revoke',
-          expect.objectContaining({
-            method: 'DELETE'
-          })
-        );
-      }
-    });
-  });
-
-  describe('Edge Cases', () => {
-    it('should handle multiple GitHub accounts for same user', async () => {
-      // Create second GitHub account for same user
-      const secondAccount = await db.account.create({
-        data: {
-          userId: testUser.id,
-          type: 'oauth',
-          provider: 'github',
-          providerAccountId: '789012',
-          access_token: 'encrypted-access-token-2'
-        }
-      });
-
-      mockGetServerSession.mockResolvedValue({
-        user: { id: testUser.id },
-        expires: '2024-12-31'
-      } as any);
-
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      const response = await POST();
-      const data = await response.json();
-
-      expect(response.status).toBe(200);
-      expect(data.success).toBe(true);
-
-      // Should only delete the first account found
-      const remainingAccounts = await db.account.findMany({
-        where: { userId: testUser.id, provider: 'github' }
-      });
-      expect(remainingAccounts).toHaveLength(1);
-
-      // Cleanup
-      await db.account.delete({
-        where: { id: secondAccount.id }
-      });
-    });
-
-    it('should handle session cleanup when sessions already deleted', async () => {
-      // Pre-delete sessions
-      await db.session.deleteMany({
-        where: { userId: testUser.id }
-      });
-
-      mockGetServerSession.mockResolvedValue({
-        user: { id: testUser.id },
-        expires: '2024-12-31'
-      } as any);
-
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      const response = await POST();
-      const data = await response.json();
-
-      // Should still succeed
-      expect(response.status).toBe(200);
-      expect(data.success).toBe(true);
-    });
-  });
-
-  describe('Security Tests', () => {
-    it('should use encryption service to decrypt tokens', async () => {
-      mockGetServerSession.mockResolvedValue({
-        user: { id: testUser.id },
-        expires: '2024-12-31'
-      } as any);
-
-      const mockDecryptField = vi.fn().mockReturnValue('decrypted-token');
-      mockEncryptionService.getInstance = vi.fn().mockReturnValue({
-        decryptField: mockDecryptField
-      });
-
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      await POST();
-
-      // Verify decryption was called if GitHub API call was made
-      if (mockFetch.mock.calls.length > 0) {
-        expect(mockDecryptField).toHaveBeenCalledWith(
-          'access_token',
-          testAccount.access_token
-        );
-      }
-    });
-
-    it('should use proper authorization header for GitHub API', async () => {
-      mockGetServerSession.mockResolvedValue({
-        user: { id: testUser.id },
-        expires: '2024-12-31'
-      } as any);
-
-      const request = new NextRequest('http://localhost:3000/api/auth/revoke-github', {
-        method: 'POST'
-      });
-
-      await POST();
-
-      // Only check if GitHub API was called (it should be since we have access token)
-      if (mockFetch.mock.calls.length > 0) {
-        const authHeader = mockFetch.mock.calls[0][1].headers.Authorization;
-        expect(authHeader).toMatch(/^Basic [A-Za-z0-9+/]+=*$/);
-        
-        // Verify it's properly base64 encoded client credentials
-        const base64Credentials = authHeader.replace('Basic ', '');
-        const credentials = Buffer.from(base64Credentials, 'base64').toString();
-        expect(credentials).toMatch(/^.+:.+$/); // Should be in format client_id:client_secret
-      }
-    });
-  });
-});
